Add files via upload
I have built a project that generates a password for you and saves it in a data file, so if you ever forget the password, you can check it there. You just need to enter the website name to remember what it’s for, and then click on Generate Password.